refactor: split seo-content-analyzer.py into focused modules (t1488)
Split 745-line seo-content-analyzer.py (Qlty file-complexity 177) into
focused single-responsibility modules:

- seo_extraction.py: HTML/content parsing utilities (141 lines)
- seo_readability.py: ReadabilityScorer class (153 lines)
- seo_keywords.py: KeywordAnalyzer class (158 lines)
- seo_intent.py: SearchIntentAnalyzer class (60 lines)
- seo_quality.py: SEOQualityRater class (160 lines)
- seo_scoring.py: compat shim (re-exports all 4 classes)
- seo_readability_keywords.py: compat shim (re-exports ReadabilityScorer, KeywordAnalyzer)
- seo_intent_quality.py: compat shim (re-exports SearchIntentAnalyzer, SEOQualityRater)
- seo-content-analyzer.py: CLI entry point (186 lines, imports via seo_scoring)

All modules pass Qlty smells check (zero file-complexity or duplication
smells). CLI interface unchanged — all commands verified working.

Closes #4908
